Slack alerts
Adding the ability to match a regex when a PR is merged to a slack
alert.  This allows you to notifiy channels when PRs are merged that
follow your CAB process or you can use it to alert on special
circumstances such as an emergency PR.

The message provided in the alerts config gets put through go templater
with the PR event so you can pick out any bits from the PR that you wish
to send to slack.  The slack webhook url is encrypted in the config
using AES 256 as that url is sensitive.