Dismiss reviews by contributors to the PR
Previous to this change, a reviewer could contribute commits to the PR
while still be able to approve the PR. With this change we remove that
ability to keep the invariant of an reviewer should not also be a
contributor.

Given a PR where Alice and Bob contributed to it and Bob also approved
the PR. The PR will then still remain in pending state until another
member of the team approves it.
The bot will post a comment on the PR tagging Bob as a dismissed
reviewer.

If a new review is pushed, the previous dismissed review comment will
be deleted and a new one is published if the review is still not
sufficient to avoid spamming comments.

Other changes:
- Turn install-pact into a real target & make it a dependency to test
  - This way we only invoke the target if the binary isn't present