Introduce structure in application
With the work 'dismissal of reviewers who have committed to the PR',
we found it difficult to modify the application due to the lack of
structure.

- related functions were not grouped in packages under internal/
- multiple variables being returned from functions instead of using
  structs
- global variables that were being manipulated in the same package
  (due to usage of packages, it became hard to understand when and
  what something happened)
- not handling SIGTERM, SIGINT, no graceful shutdown
- ...

We haven't solved all of the problems we've observed, however we've
introduced boundaries, components which will make it easier to work
with the repository.