feat: regex, regex_label and directories in pull_request_approval_rule now a logical "and" between them.
In order to classify a pull request approval rule as matching, pull request must pass ALL conditions in a rule.
If regex, regex_label and directores are not empty, they all have to match.
If one of these fields is empty, it will not be considered.

BREAKING CHANGE: regex, regex_label and directories now have logical "and" between them.